async cargo test, highlight cword and line excess
Hi, thanks so much for all the videos and sharing your setup, it is amazing.
I have a couple of potentially handy tricks that could be useful for you as well:
I use a secondary match (2match) to highlight the excess characters after a number of chars, I find the color column confusing because I tend to have several window open at the same time.
The second one is that I'm a walking typo and I really like to see that the variable I typed is the same as the variable three lines above that I intend to reference, I have updatetime=200 to lower the time it would be highlighted.
Lastly, I use the the async jobstart to run cargo test (`<leader> tc`) and display the test status (OK/Err) in my lightline, on errors it uses a temporary buffer split to show the test errors, not sure you would find this useful.
Either way, thanks again for the videos! They are really amazing!